src/test/java/seedu/duke/commands/member/TeamMemberHoursCommandTest.java
package seedu.duke.commands.member;

import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.Test;
import seedu.duke.DukeExceptions;
import seedu.duke.member.TeamMember;
import seedu.duke.project.Project;
import seedu.duke.task.Task;
import seedu.duke.ui.Ui;

import java.util.ArrayList;
import java.util.HashMap;

import static org.junit.jupiter.api.Assertions.*;

class TeamMemberHoursCommandTest {
    static ArrayList<Project> projects;
    static ArrayList<TeamMember> teamMembers = new ArrayList<>();
    static Ui ui = new Ui();

    @BeforeAll
    static void createMemberList() {
        TeamMember member1 = new TeamMember("Member 1");
        TeamMember member2 = new TeamMember("Member 2");
        TeamMember member3 = new TeamMember("Member 3");

        teamMembers.add(member1);
        teamMembers.add(member2);
        teamMembers.add(member3);
    }

    @Test
    public void executeCommand_validMemberId_correctHours() throws DukeExceptions {
        Task task1 = new Task("task1");
        task1.addActual(274);

        Task task2 = new Task("task2");
        task2.addActual(128);

        Task task3 = new Task("task3");
        task3.addActual(45);

        Task task4 = new Task("task4");
        task4.addActual(290);

        TeamMember member3 = teamMembers.get(2);
        member3.setTask(task1);
        member3.setTask(task2);
        member3.setTask(task3);
        member3.setTask(task4);

        HashMap<String, String> params = new HashMap<>();
        params.put("m","3");

        TeamMemberHoursCommand assignProject = new TeamMemberHoursCommand(params,1);
        String expectedOutput = "Member 3 worked for 12.3 hours.";
        String actualOutput = assignProject.executeCommand(projects,teamMembers);
        assertEquals(expectedOutput, actualOutput);
    }

    @Test
    public void executeCommand_invalidMemberId_assignMemberToProject() throws DukeExceptions {
        HashMap<String,String> params = new HashMap<>();
        params.put("m","7");

        TeamMemberHoursCommand assignProject = new TeamMemberHoursCommand(params,1);
        String expectedOutput = "The member ID entered does not exist";
        Throwable actualOutputException = assertThrows(DukeExceptions.class, () -> {
            assignProject.executeCommand(projects,teamMembers);
        });
        assertEquals(expectedOutput,actualOutputException.toString());
    }
}